Add `From<PanicReason>` for io::Error
We often need to convert directly from a panic reason to an io error,
such as when performing to/from bytes encoding - in that case, the VM
will not be intermediary to the operation and the conversion will be
performed directly.

This commit introduces a transparent conversion between the error types.